Move prependThemeFilterKeys
