+++ client/my-sites/themes/themes-selection.jsx
@@ -25,10 +25,10 @@ import {
 	isThemesLastPageForQuery,
 	isThemeActive,
 	isInstallingTheme,
+	prependThemeFilterKeys,
 } from 'state/themes/selectors';
 import { setThemePreviewOptions } from 'state/themes/actions';
 import config from 'config';
-import prependThemeFilterKeys from 'state/selectors/prepend-theme-filter-keys';
 
 /**
  * Style dependencies
