+++ client/state/themes/selectors/index.js
@@ -52,6 +52,7 @@ export { isThemesLastPageForQuery } from 'state/themes/selectors/is-themes-last-
 export { isValidThemeFilterTerm } from 'state/themes/selectors/is-valid-theme-filter-term';
 export { isWpcomTheme } from 'state/themes/selectors/is-wpcom-theme';
 export { isWporgTheme } from 'state/themes/selectors/is-wporg-theme';
+export { prependThemeFilterKeys } from 'state/themes/selectors/prepend-theme-filter-keys';
 export { shouldFilterWpcomThemes } from 'state/themes/selectors/should-filter-wpcom-themes';
 export { shouldShowHomepageWarning } from 'state/themes/selectors/should-show-homepage-warning';
 export { themeHasAutoLoadingHomepage } from 'state/themes/selectors/theme-has-auto-loading-homepage';
