+++ client/state/themes/selectors/prepend-theme-filter-keys.js
@@ -1,7 +1,9 @@
 /**
  * Internal dependencies
  */
-import { getThemeFilterStringFromTerm } from 'state/themes/selectors';
+import { getThemeFilterStringFromTerm } from 'state/themes/selectors/get-theme-filter-string-from-term';
+
+import 'state/themes/init';
 
 /**
  * For a string of terms, recreate full search string in
@@ -12,7 +14,7 @@ import { getThemeFilterStringFromTerm } from 'state/themes/selectors';
  * @param {string} terms Space or + separated list of filter terms
  * @returns {string} Complete taxonomy:term filter string, or empty string if term is not valid
  */
-export default function prependThemeFilterKeys( state, terms = '' ) {
+export function prependThemeFilterKeys( state, terms = '' ) {
 	const result = terms
 		.split( /[+\s]/ )
 		.map( term => getThemeFilterStringFromTerm( state, term ) )
